Split mill src by concern + one file per operator
operators.ts was ~750 lines. Extract the shared machinery into cohesive
modules and give each operator its own file:

- undo.ts        — MutableUndo + undo helpers + capturePathUndo
- array-ops.ts   — in-place array primitives (push/remove/sort/slice/parse)
- operators/shared.ts — OperatorContext, eachPath, writeNumeric, removeByPredicate
- operators/{set,unset,inc,mul,min,max,rename,push,pop,pull,pull-all,add-to-set}.ts
- operators.ts   — operation types + dispatch table + update() entry point

Pure reorganization: no behavior change, 189 tests pass, 100% coverage.
