+++ packages/mill/src/operators/mul.ts
@@ -1,0 +1,10 @@
+import { eachPath, type OperatorContext, writeNumeric } from "./shared";
+
+export function $mul(context: OperatorContext, operations: Record<string, number>): void {
+  eachPath(context, operations as Record<string, unknown>, (path, value) => {
+    writeNumeric(context, path, {
+      operator: "$mul",
+      compute: (previous) => (typeof previous === "number" ? previous : 0) * (value as number),
+    });
+  });
+}
